requirements.txt
django>=5.2
djangorestframework>=3.15
django-cors-headers>=4.3
streamlit>=1.26
pymongo>=4.6
requests>=2.31
plotly>=5.22
pandas>=2.2